% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool1.R
\name{cells_from_mc_hm}
\alias{cells_from_mc_hm}
\title{Get metacells names from the heatmap of the metacells object}
\usage{
cells_from_mc_hm(
  data = NULL,
  metadata = NULL,
  HMkc = 40,
  HMkr = 0,
  rmclu = NULL,
  name = NULL
)
}
\arguments{
\item{data}{Metacell z matrix}

\item{metadata}{metadata with cells corresponding to metacells}

\item{HMkc}{Number of k groups of columns}

\item{HMkr}{Number of k groups of rows}

\item{rmclu}{list of clusters to remove, ex : "1,2,4"}

\item{name}{Name of the file}
}
\value{
A matrix of the infile
}
\description{
Get metacells names from the heatmap of the metacells object
}
