% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool1.R
\name{vm_genes_in_mod}
\alias{vm_genes_in_mod}
\title{Get the gene to use in module analysis and plot them
They are the most variable+expressed genes, following loess curve}
\usage{
vm_genes_in_mod(
  varmean_df = NULL,
  vm_MeanThres = -1,
  vm_VarmeanThres = 1,
  x1min = -1,
  x1max = 3.5
)
}
\arguments{
\item{varmean_df}{dataframe of mean and variance of the genes}

\item{vm_MeanThres}{threshold for the log(mean) of expression}

\item{vm_VarmeanThres}{threshold for the curve start}

\item{x1min}{min value for the plot}

\item{x1max}{max value for the plot}
}
\value{
TRUE/FALSE genes kept
}
\description{
Get the gene to use in module analysis and plot them
They are the most variable+expressed genes, following loess curve
}
